+++ spec/appydays/loggable_spec.rb
@@ -197,476 +197,4 @@ RSpec.describe Appydays::Loggable do
       SemanticLogger.remove_appender(other_appender)
     end
   end
-
-  describe Appydays::Loggable::RequestLogger do
-    def run_app(app, opts: {}, loggers: [], env: {}, cls: Appydays::Loggable::RequestLogger)
-      rl = cls.new(app, **opts, reraise: false)
-      return capture_logs_from(loggers << rl.logger, formatter: :json) do
-        _, _, body = rl.call(env)
-        body&.close
-      end
-    end
-
-    it "logs info about the request" do
-      lines = run_app(proc { [200, {}, ""] })
-      expect(lines).to have_a_line_matching(/"message":"request_finished".*"response_status":200/)
-
-      lines = run_app(proc { [400, {}, ""] })
-      expect(lines).to have_a_line_matching(/"message":"request_finished".*"response_status":400/)
-    end
-
-    it "logs content length" do
-      lines = run_app(proc { [200, {"content-LENGTH" => "5"}, ""] })
-      expect(lines).to have_a_line_matching(/"response_content_length":"5"/)
-    end
-
-    it "logs at 599 (or configured value) if something errors" do
-      lines = run_app(proc { raise "testing error" })
-      expect(lines).to have_a_line_matching(/"level":"error".*"response_status":599/)
-      expect(lines).to have_a_line_matching(/"message":"testing error"/)
-    end
-
-    it "logs slow queries at warn" do
-      lines = run_app(proc { [200, {}, ""] }, opts: {slow_request_seconds: 0})
-      expect(lines).to have_a_line_matching(/"level":"warn".*"response_status":200/)
-    end
-
-    it "logs errors at error" do
-      lines = run_app(proc { [504, {}, ""] }, opts: {slow_request_seconds: 0})
-      expect(lines).to have_a_line_matching(/"level":"error".*"response_status":504/)
-    end
-
-    it "adds request_id tag around the execution of the request" do
-      logger = SemanticLogger["testlogger"]
-      lines = run_app(proc do
-                        logger.info("check for tags")
-                        [200, {}, ""]
-                      end,
-                      opts: {slow_request_seconds: 0}, loggers: [logger],)
-      expect(lines).to have_attributes(length: 2)
-      # This should have request_id, but not the other request tags, like the path
-      expect(lines[0]).to include_json(message: "check for tags", context: include("request_id"))
-      expect(lines[0]).to include_json(context: not_include("request_path"))
-      # This should have all tags
-      expect(lines[1]).to include_json(message: "request_finished", context: include("request_id", "request_path"))
-    end
-
-    it "adds subclass tags" do
-      ReqLogger = Class.new(Appydays::Loggable::RequestLogger) do
-        def request_tags(env)
-          return {my_header_tag: env["HTTP_MY_HEADER"]}
-        end
-      end
-      lines = run_app(proc { [200, {}, ""] }, env: {"HTTP_MY_HEADER" => "myval"}, cls: ReqLogger)
-      expect(lines).to have_a_line_matching(/"my_header_tag":"myval"/)
-    end
-
-    it "adds and sets request and trace ids if trace and request headers not set" do
-      env = {}
-      trace_id = nil
-      request_id = nil
-      lines = run_app(proc do
-        trace_id = env.fetch("HTTP_TRACE_ID")
-        request_id = env.fetch("HTTP_X_REQUEST_ID")
-        [200, {}, ""]
-      end, env:,)
-      expect(lines).to contain_exactly(
-        include_json(message: "request_finished", context: include(
-          "trace_id" => trace_id,
-          "request_id" => request_id,
-        ),),
-      )
-    end
-
-    it "reads the request and trace id headers" do
-      env = {"HTTP_TRACE_ID" => "t1", "HTTP_X_REQUEST_ID" => "r1"}
-      lines = run_app(proc do
-        expect(env).to include("HTTP_TRACE_ID" => "t1", "HTTP_X_REQUEST_ID" => "r1")
-        [200, {}, ""]
-      end, env:,)
-      expect(lines).to contain_exactly(
-        include_json(message: "request_finished", context: include(
-          "trace_id" => "t1",
-          "request_id" => "r1",
-        ),),
-      )
-    end
-
-    it "will use the trace id header" do
-      env = {"HTTP_TRACE_ID" => "t1"}
-      lines = run_app(proc do
-        expect(env).to include("HTTP_TRACE_ID" => "t1", "HTTP_X_REQUEST_ID" => have_attributes(length: 36))
-        [200, {}, ""]
-      end, env:,)
-      expect(lines).to contain_exactly(
-        include_json(message: "request_finished", context: include(
-          "trace_id" => "t1",
-          "request_id" => have_attributes(length: 36),
-        ),),
-      )
-    end
-
-    it "can add log fields to the request_finished message" do
-      lines = run_app(proc do
-        described_class.set_request_tags(abc: "123")
-        [200, {}, ""]
-      end)
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "info",
-          name: "Appydays::Loggable::RequestLogger",
-          message: "request_finished",
-          context: include("abc" => "123"),
-        ),
-      )
-    end
-
-    it "clears tags after the request (even if it raises an error)" do
-      lines = run_app(proc do
-        described_class.set_request_tags(tag1: "a")
-        [200, {}, ""]
-      end)
-      expect(lines).to contain_exactly(include_json(context: include("tag1" => "a")))
-
-      lines = run_app(proc do
-        described_class.set_request_tags(tag2: "a")
-        raise "oops"
-      end)
-      expect(lines).to contain_exactly(include_json(context: include("tag2" => "a")))
-      expect(lines).to contain_exactly(include_json(context: not_include("tag1" => "a")))
-
-      lines = run_app(proc do
-        described_class.set_request_tags(tag3: "a")
-        [400, {}, ""]
-      end)
-      expect(lines).to contain_exactly(include_json(context: include("tag3" => "a")))
-      expect(lines).to contain_exactly(include_json(context: not_include("tag1" => "a")))
-      expect(lines).to contain_exactly(include_json(context: not_include("tag2" => "a")))
-    end
-  end
-
-  describe Appydays::Loggable::SidekiqJobLogger, :db do
-    before(:each) { @slow_secs = 5 }
-    let(:logcls) do
-      slow_secs = @slow_secs
-      Class.new(Appydays::Loggable::SidekiqJobLogger) do
-        attr_accessor :slow_secs
-
-        def self.name
-          return "TestLogger"
-        end
-        define_method(:slow_job_seconds) { slow_secs }
-      end
-    end
-
-    let(:logger) { logcls.new(Sidekiq::Config.new) }
-
-    def log(reraise: true, &block)
-      block ||= proc {}
-      lines = capture_logs_from(logcls.logger, formatter: :json) do
-        logger.call({}, nil) do
-          block.call
-        end
-      rescue StandardError => e
-        raise e if reraise
-      end
-      return lines
-    end
-
-    it "logs a info message for the job" do
-      lines = log
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "info",
-          name: "TestLogger",
-          message: "job_done",
-          duration_ms: be_a(Numeric),
-        ),
-      )
-    end
-
-    it "logs duration properly (SemanticLogger uses milliseconds)" do
-      lines = log { sleep(0.001) }
-      expect(lines).to contain_exactly(
-        include_json(
-          message: "job_done",
-          duration: match(/^\d+\.\d+ms$/),
-          duration_ms: be >= 1,
-        ),
-      )
-    end
-
-    it "logs at warn if the time taken is more than the slow job seconds" do
-      @slow_secs = 0
-      lines = log
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "warn",
-          name: "TestLogger",
-          message: "job_done",
-          duration_ms: be_a(Numeric),
-        ),
-      )
-    end
-
-    it "logs at error (but does not log the exception) if the job fails" do
-      lines = log(reraise: false) do
-        1 / 0
-      end
-
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "error",
-          name: "TestLogger",
-          message: "job_fail",
-          duration_ms: be_a(Numeric),
-        ),
-      )
-      expect(lines[0]).to_not include("exception")
-    end
-
-    it "can add log fields to the job_done message" do
-      lines = log do
-        described_class.set_job_tags(tag1: "hi")
-      end
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "info",
-          name: "TestLogger",
-          message: "job_done",
-          context: include("tag1" => "hi"),
-        ),
-      )
-    end
-
-    it "can add log fields to the job_fail message" do
-      lines = log(reraise: false) do
-        described_class.set_job_tags(tag1: "hi")
-        raise "hello"
-      end
-      expect(lines).to contain_exactly(
-        include_json(
-          level: "error",
-          name: "TestLogger",
-          message: "job_fail",
-          context: include("tag1" => "hi"),
-        ),
-      )
-    end
-
-    it "clears job tags after the job" do
-      lines = log do
-        described_class.set_job_tags(tag1: "hi")
-      end
-      expect(lines).to contain_exactly(
-        include_json(
-          context: include("tag1" => "hi"),
-        ),
-      )
-
-      lines = log
-      expect(lines).to contain_exactly(
-        include_json(
-          context: not_include("tag1"),
-        ),
-      )
-    end
-  end
-
-  describe "Sequel Logging" do
-    after(:each) do
-      Sequel::Database::AppydaysLogger.setdefaults
-    end
-    describe "with structure logging" do
-      def log
-        logger = SemanticLogger[Sequel]
-        db = Sequel.connect("mock://", logger:, log_warn_duration: 3)
-        return capture_logs_from(logger, formatter: :json) do
-          yield(db)
-        end
-      end
-
-      it "logs info" do
-        lines = log do |db|
-          db.log_info("hello1")
-          db.log_info("hello2", {x: 1})
-        end
-        expect(lines).to contain_exactly(
-          include_json(
-            level: "info",
-            name: "Sequel",
-            message: "sequel_log",
-            context: {"message" => "hello1"},
-          ),
-          include_json(
-            context: {"message" => "hello2", "args" => {"x" => 1}},
-          ),
-        )
-      end
-
-      it "logs exceptions" do
-        lines = log do |db|
-          db.log_exception(RuntimeError.new("nope"), "msg")
-        end
-        expect(lines).to contain_exactly(
-          include_json(
-            level: "error",
-            message: "sequel_exception",
-            exception: {"name" => "RuntimeError", "message" => "nope", "stack_trace" => nil},
-            context: {"sequel_message" => "msg"},
-          ),
-        )
-      end
-
-      it "logs 'table exists' exceptions at debug" do
-        lines = log do |db|
-          db.log_exception(RuntimeError.new("nope"), "SELECT NULL AS nil FROM sch.foobar LIMIT 1")
-          db.log_exception(RuntimeError.new("nope"), "select null as \"nil\" FROM \"sch\".\"foobar\" LIMIT 1")
-        end
-        expect(lines).to contain_exactly(
-          include_json(level: "debug", message: "sequel_exception"),
-          include_json(level: "debug", message: "sequel_exception"),
-        )
-      end
-
-      it "logs duration" do
-        lines = log do |db|
-          db.log_duration(4, "slow")
-          db.log_duration(1, "fast")
-        end
-        expect(lines).to contain_exactly(
-          include_json(
-            level: "warn",
-            message: "sequel_query",
-            duration_ms: 4000,
-            duration: "4.000s",
-            context: {"query" => "slow"},
-          ),
-          include_json(
-            level: "info",
-            message: "sequel_query",
-            context: {"query" => "fast"},
-          ),
-        )
-      end
-
-      it "truncates long messages at debug" do
-        Sequel::Database::AppydaysLogger.truncation_context = 3
-        Sequel::Database::AppydaysLogger.truncation_message = "<truncated>"
-        Sequel::Database::AppydaysLogger.log_full_message_level = :debug
-
-        lines = log do |db|
-          db.log_duration(4, "a" * 3000)
-          db.log_duration(1, "a" * 3000)
-        end
-
-        expect(lines).to contain_exactly(
-          include_json(
-            level: "warn",
-            message: "sequel_query",
-            context: {"query" => "aaa<truncated>aaa", "truncated" => true},
-          ),
-          include_json(
-            level: "debug",
-            message: "sequel_query_debug",
-            context: {"query" => have_length(3000)},
-          ),
-          include_json(
-            level: "info",
-            message: "sequel_query",
-            context: {"query" => "aaa<truncated>aaa", "truncated" => true},
-          ),
-          include_json(
-            level: "debug",
-            message: "sequel_query_debug",
-            context: {"query" => have_length(3000)},
-          ),
-        )
-      end
-
-      it "does not log untruncated messages if log_full_message_level is nil" do
-        Sequel::Database::AppydaysLogger.truncation_context = 3
-        Sequel::Database::AppydaysLogger.truncation_message = "<truncated>"
-        Sequel::Database::AppydaysLogger.log_full_message_level = nil
-
-        lines = log do |db|
-          db.log_duration(4, "a" * 3000)
-          db.log_duration(1, "a" * 3000)
-        end
-
-        expect(lines).to contain_exactly(
-          include_json(
-            level: "warn",
-            message: "sequel_query",
-            context: {"query" => "aaa<truncated>aaa", "truncated" => true},
-          ),
-          include_json(
-            level: "info",
-            message: "sequel_query",
-            context: {"query" => "aaa<truncated>aaa", "truncated" => true},
-          ),
-        )
-      end
-    end
-
-    describe "with standard logging" do
-      def log
-        device = StringIO.new
-        logger = Logger.new(device)
-        db = Sequel.connect("mock://", logger:, log_warn_duration: 3)
-        yield(db)
-        return device.string
-      end
-
-      it "logs info" do
-        lines = log do |db|
-          db.log_info("hello1")
-          db.log_info("hello2", {x: 1})
-        end
-        tags = "{x: 1}"
-        tags = "{:x=>1}" if Gem::Version.new(RUBY_VERSION) < Gem::Version.new("3.4.0")
-        expect(lines.lines).to contain_exactly(
-          include("INFO -- : hello1\n"),
-          include("INFO -- : hello2; #{tags}"),
-        )
-      end
-
-      it "logs exceptions" do
-        lines = log do |db|
-          db.log_exception(RuntimeError.new("nope"), "msg")
-        end
-        expect(lines.lines).to contain_exactly(
-          include("ERROR -- : RuntimeError: nope: msg\n"),
-        )
-      end
-
-      it "logs duration" do
-        lines = log do |db|
-          db.log_duration(4, "slow")
-          db.log_duration(1, "fast")
-        end
-        expect(lines.lines).to contain_exactly(
-          include("INFO -- : (1.000000s) fast\n"),
-          include("WARN -- : (4.000000s) slow\n"),
-        )
-      end
-    end
-  end
-
-  describe "HTTParty formatter" do
-    it "logs structured request information" do
-      logger = SemanticLogger["http_spec_logging_test"]
-      stub_request(:post, "https://foo/bar").to_return(status: 200, body: "")
-      logs = capture_logs_from(logger, formatter: :json) do
-        HTTParty.post("https://foo/bar", body: {x: 1}, logger:, log_format: :appydays)
-      end
-      expect(logs).to contain_exactly(
-        include_json(
-          "message" => "httparty_request",
-          "level" => "info",
-          "context" => include("http_method" => "POST"),
-        ),
-      )
-    end
-  end
 end
